fix(telegram): return JSON body from webhook handler + fix secret token sync
Two bugs causing "not responding" symptom:

1. webhook_handler returned bare `StatusCode` — axum serializes this with
   no body, so the proxy and Telegram both saw empty responses on every path.
   Changed return type to `(StatusCode, Json<Value>)` with ok_body()/err_body()
   helpers returning {"ok": true} / {"ok": false, "error": "..."}.

2. `trusty-telegram webhook set` was calling api_set_webhook(..., None),
   registering the URL without the stored secret token. Subsequent Telegram
   updates included no X-Telegram-Bot-Api-Secret-Token header → 403 on every
   request (logged at warn, body was empty so caller saw nothing).
   Fix: CLI Set command now reads webhook_secret_token from SQLite and passes
   it to Telegram so secrets stay in sync. Rejection log upgraded to error!
   with diagnostic context and a remediation hint.

Also fix clippy: remove redundant `as usize` cast in trusty-store/src/lance.rs
and use `if let Some` instead of `.is_some()` + `.unwrap()` in engine.rs.
cargo fmt applied to telegram_push.rs.

+++ crates/trusty-store/src/lance.rs
@@ -128,10 +128,7 @@ async fn open_or_create(
         Err(_) => {
             // Table doesn't exist — create it empty
             let empty_batch = RecordBatch::new_empty(schema.clone());
-            let tbl = conn
-                .create_table(name, vec![empty_batch])
-                .execute()
-                .await?;
+            let tbl = conn.create_table(name, vec![empty_batch]).execute().await?;
             Ok(tbl)
         }
     }
@@ -477,7 +474,7 @@ impl LanceStore {
         let ql = query.to_lowercase();
 
         // Use actual row count to avoid "Ran out of fragments" warning.
-        let row_count = table.count_rows(None).await.unwrap_or(20_000) as usize;
+        let row_count = table.count_rows(None).await.unwrap_or(20_000);
         let stream = table.query().limit(row_count).execute().await?;
         let batches: Vec<RecordBatch> = collect_stream(stream).await?;
         let mut matched = Vec::new();
